requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.2
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
tabulate==0.8.5
urllib3==1.25.6
